+++ src/rust/src/lib.rs
@@ -22,6 +22,7 @@ pub use r_bindings::r_enrichment::r_mitch;
 pub use r_bindings::r_enrichment::r_oea;
 
 // graphs
+pub use r_bindings::r_graph::r_graph_clustering;
 pub use r_bindings::r_graph::r_page_rank;
 pub use r_bindings::r_graph::r_snf;
 
@@ -55,6 +56,7 @@ extendr_module! {
     // graphs
     use r_page_rank;
     use r_snf;
+    use r_graph_clustering;
     // methods
     use r_coremo;
     use r_dgrdl;